pyproject.toml
[project]
name = "espo-implementation-tool"
version = "0.1.0"
description = "Desktop tool for deploying EspoCRM field configurations from YAML"
readme = "README.md"
authors = [{ name = "Doug", email = "doug@dougbower.com" }]
requires-python = ">=3.12"
dependencies = [
    "pyside6>=6.10.2",
    "requests>=2.32.5",
    "pyyaml>=6.0.2",
    "python-docx>=1.2.0",
]

[project.scripts]
espo-impl = "espo_impl.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["espo_impl"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "--strict-config"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
